chore: rename data table
